refactor(node/fs): support 'Symbol.for(nodejs.util.promisify.custom)'
